+++ src/core/trial.cpp
@@ -73,11 +73,21 @@ bool Trial::init()
         }
     }
 
+    bool ok = false;
+    auto edgeAttrsGen = m_exp->edgeAttrsGen(ok);
+    if (!ok) {
+        qWarning() << "unable to create the trials."
+                   << "Failed to create the edges' attributes generator.\n"
+                   << "Experiment:" << m_exp->id();
+        return false;
+    }
+
     const quint32 seed = m_exp->inputs()->general(GENERAL_ATTR_SEED).toUInt();
     m_prg = new PRG(seed + m_id);
 
     m_graph = dynamic_cast<AbstractGraph*>(m_exp->graphPlugin()->create());
-    if (!m_graph || !m_graph->setup(*this, *m_exp->inputs()->graph(), nodes)) {
+    if (!m_graph || !m_graph->setup(*this, std::move(edgeAttrsGen),
+                                    *m_exp->inputs()->graph(), nodes)) {
         qWarning() << "unable to create the trials."
                    << "The graph could not be initialized."
                    << "Experiment:" << m_exp->id();